requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
google-generativeai==0.8.3
python-dotenv==1.0.0
gunicorn==21.1.0
Werkzeug==3.0.1
